+++ client/planner-app/index.js
@@ -12,7 +12,6 @@ page('/', utils.redirect('/planner'));
 page('/login', require('login-page'));
 page('/forgot-password', require('forgot-password-page'));
 page('/change-password/:key', require('change-password-page'));
-page('/welcome', session.commuterIsLoggedIn, Plan.load, require('welcome-page'));
 page('/planner', session.commuterIsLoggedIn, Plan.load, require('planner-page'));
 page('/planner/:link', session.loginWithLink);